release - This is a copy of OpenLMS Production45 code

+++ local/o365/export_manifest.php
@@ -27,11 +27,17 @@ require_once(__DIR__ . '/../../config.php');
 require_once($CFG->libdir . '/filestorage/zip_archive.php');
 require_once($CFG->dirroot . '/local/o365/lib.php');
 
+require_admin();
+
 // Mark manifest file as downloaded.
+$existingmanifestdownloadedsetting = get_config('local_o365', 'manifest_downloaded');
+if (!$existingmanifestdownloadedsetting) {
+    add_to_config_log('manifest_downloaded', $existingmanifestdownloadedsetting, true, 'local_o365');
+}
 set_config('manifest_downloaded', true, 'local_o365');
 purge_all_caches();
 
-list($error, $manifestfilepath) = local_o365_create_manifest_file();
+[$errorcode, $manifestfilepath] = local_o365_create_manifest_file();
 
 if ($manifestfilepath) {
     // Download manifest file.
@@ -42,5 +48,5 @@ if ($manifestfilepath) {
     header("Expires: 0");
     readfile($manifestfilepath);
 } else {
-    print_error($error);
+    throw new moodle_exception($errorcode, 'local_o365');
 }
